+++ libs/matdbg/src/DebugServer.cpp
@@ -24,6 +24,7 @@
 #include <spirv_glsl.hpp>
 #include <spirv-tools/libspirv.h>
 
+#include <matdbg/ShaderReplacer.h>
 #include <matdbg/ShaderExtractor.h>
 #include <matdbg/ShaderInfo.h>
 #include <matdbg/JsonWriter.h>
@@ -32,6 +33,7 @@
 
 #include <backend/DriverEnums.h>
 
+#include <sstream>
 #include <string>
 
 // If set to 0, this serves HTML from a resgen resource. Use 1 only during local development, which
@@ -202,13 +204,6 @@ public:
         }
 
         if (glindex[0]) {
-            ShaderExtractor extractor(Backend::OPENGL, result->package, result->packageSize);
-            if (!extractor.parse() ||
-                    (!extractor.isShadingMaterial() && !extractor.isPostProcessMaterial())) {
-                return error(__LINE__);
-            }
-
-            filaflat::ShaderBuilder builder;
             std::vector<ShaderInfo> info(getShaderCount(package, ChunkType::MaterialGlsl));
             if (!getGlShaderInfo(package, info.data())) {
                 return error(__LINE__);
@@ -219,7 +214,14 @@ public:
                 return error(__LINE__);
             }
 
+            ShaderExtractor extractor(Backend::OPENGL, result->package, result->packageSize);
+            if (!extractor.parse() ||
+                    (!extractor.isShadingMaterial() && !extractor.isPostProcessMaterial())) {
+                return error(__LINE__);
+            }
+
             const auto& item = info[shaderIndex];
+            filaflat::ShaderBuilder builder;
             extractor.getShader(item.shaderModel, item.variant, item.pipelineStage, builder);
 
             mg_printf(conn, kSuccessHeader.c_str(), "application/txt");
@@ -316,7 +318,41 @@ public:
 
     bool handleData(CivetServer *server, struct mg_connection *conn, int bits, char *data,
             size_t size) override {
-        // TODO: trigger the "shader has been edited" callback via mServer->mEditHandler(...)
+        // TODO: Is there a better way to ignore the handshake message that occurs after startup?
+        if  (size < 8) {
+            return true;
+        }
+
+        // Every WebSocket message is prefixed with a 4-character command followed by a space.
+        //
+        // For now we simply use istringstream for parsing, so command arguments are delimited
+        // with space characters.
+        //
+        // The "API index" matches the values of filament::backend::Backend (zero is invalid).
+        //
+        // The "shader index" is a zero-based index into the list of variants using the order that
+        // they appear in the package, where each API (GL / VK / Metal) has its own list.
+        //
+        // Commands:
+        //
+        //     EDIT [material id] [api index] [shader index] [entire shader source....]
+        //
+
+        const static StaticString kEditCmd = "EDIT ";
+        const static size_t kEditCmdLength = kEditCmd.size();
+
+        if (strncmp(data, kEditCmd.c_str(), kEditCmdLength)) {
+            slog.e << "Bad WebSocket message." << io::endl;
+            return false;
+        }
+        std::istringstream str(data + kEditCmdLength);
+        uint32_t matid;
+        int api;
+        int shaderIndex;
+        str >> std::hex >> matid >> std::dec >> api >> shaderIndex;
+        const char* source = data + kEditCmdLength + str.tellg();
+        const size_t remaining = size - kEditCmdLength - str.tellg();
+        mServer->handleEditCommand(matid, backend::Backend(api), shaderIndex, source, remaining);
         return true;
     }
 
@@ -408,6 +444,76 @@ void DebugServer::addMaterial(const CString& name, const void* data, size_t size
 const DebugServer::MaterialRecord* DebugServer::getRecord(const MaterialKey& key) const {
     const auto& iter = mMaterialRecords.find(key);
     return iter == mMaterialRecords.end() ? nullptr : &iter->second;
+}
+
+bool DebugServer::handleEditCommand(const MaterialKey& key, backend::Backend api, int shaderIndex,
+            const char* source, size_t size) {
+    const auto error = [](int line) {
+        slog.e << "DebugServer: Unable to apply shader edit at line " << line << io::endl;
+        return false;
+    };
+
+    if (mMaterialRecords.find(key) == mMaterialRecords.end()) {
+        return error(__LINE__);
+    }
+    MaterialRecord& material = mMaterialRecords[key];
+    filaflat::ChunkContainer package(material.package, material.packageSize);
+    if (!package.parse()) {
+        return error(__LINE__);
+    }
+
+    const backend::Backend apiType = (backend::Backend) api;
+    std::vector<ShaderInfo> infos;
+    size_t shaderCount;
+    switch (apiType) {
+        case backend::Backend::OPENGL: {
+            shaderCount = getShaderCount(package, ChunkType::MaterialGlsl);
+            infos.resize(shaderCount);
+            if (!getGlShaderInfo(package, infos.data())) {
+                return error(__LINE__);
+            }
+            break;
+        }
+        case backend::Backend::VULKAN: {
+            shaderCount = getShaderCount(package, ChunkType::MaterialSpirv);
+            infos.resize(shaderCount);
+            if (!getVkShaderInfo(package, infos.data())) {
+                return error(__LINE__);
+            }
+            break;
+        }
+        case backend::Backend::METAL: {
+            shaderCount = getShaderCount(package, ChunkType::MaterialMetal);
+            infos.resize(shaderCount);
+            if (!getMetalShaderInfo(package, infos.data())) {
+                return error(__LINE__);
+            }
+            break;
+        }
+        default:
+            error(__LINE__);
+    }
+
+    if (shaderIndex < 0 || shaderIndex >= shaderCount) {
+        return error(__LINE__);
+    }
+
+    const ShaderInfo info = infos[shaderIndex];
+    ShaderReplacer editor(apiType, package.getData(), package.getSize());
+    if (!editor.replaceShaderSource(info.shaderModel, info.variant, info.pipelineStage, source, size)) {
+        return error(__LINE__);
+    }
+
+    delete [] material.package;
+
+    material.package = editor.getEditedPackage();
+    material.packageSize = editor.getEditedSize();
+
+    if (mEditCallback) {
+        mEditCallback(material.userdata, material.name, material.package, material.packageSize);
+    }
+
+    return true;
 }
 
 } // namespace matdbg